Redshift refactored and docs added
